Add REUSE.toml files and missing licenses
REUSE.toml files are read by reuse to complement or override the
copyright and licensing information found in file.

The use of REUSE.toml files was introduced in REUSE version 3.1.0a1.
This reuse version is compatible with reuse specification
version 3.2 [1].
With this commit's files,
* The SPDX document generated by reuse spdx conforms to SPDX 2.3,
* The reuse lint command reports that the Qt project is reuse compliant.

In order to be reuse compliant all the licenses referenced in file
or within a REUSE.toml files must be present in the LICENSES
directory at the base of the module.
The missing license is added.

[1]: https://reuse.software/spec-3.2/

Task-number: QTBUG-124453
Task-number: QTBUG-125211
Change-Id: Ie5215b5a691590c8bdb5ab873ac92fd2e72bfbf4
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit 99e878a198f886fe2cb222a38d60f41b97dbcc70)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>